+++ src/AdoNet/Orleans.AdoNet.Core/RelationalStorageExtensions.cs
@@ -126,7 +126,7 @@ internal static class RelationalStorageExtensions
     /// <param name="parameterProvider"></param>
     /// <typeparam name="TResult"></typeparam>
     /// <returns></returns>
-    public static Task<IEnumerable<TResult>> ReadAsync<TResult>(this IRelationalStorage storage, string query, Func<IDataRecord, TResult> selector, Action<IDbCommand> parameterProvider) => storage.ReadAsync(query, parameterProvider, (record, i, cancellationToken) => Task.FromResult(selector(record)));
+    public static Task<IEnumerable<TResult>> ReadAsync<TResult>(this IRelationalStorage storage, string query, Func<IDataRecord, TResult> selector, Action<IDbCommand>? parameterProvider) => storage.ReadAsync(query, parameterProvider, (record, i, cancellationToken) => Task.FromResult(selector(record)));
 
     /// <summary>
     /// Uses <see cref="IRelationalStorage"/> with <see cref="DbExtensions.ReflectionParameterProvider{T}(IDbCommand, T, IReadOnlyDictionary{string, string})"/>.
