+++ src/AdoNet/Orleans.Streaming.AdoNet/AdoNetStreamFailureHandler.cs
@@ -3,7 +3,7 @@ namespace Orleans.Streaming.AdoNet;
 /// <summary>
 /// An <see cref="IStreamFailureHandler"/> that attempts to move the message to dead letters.
 /// </summary>
-internal partial class AdoNetStreamFailureHandler(bool faultOnFailure, AdoNetStreamOptions streamOptions, ClusterOptions clusterOptions, AdoNetStreamQueueMapper mapper, RelationalOrleansQueries queries, ILogger<AdoNetStreamFailureHandler> logger) : IStreamFailureHandler
+internal partial class AdoNetStreamFailureHandler(bool faultOnFailure, AdoNetStreamOptions streamOptions, ClusterOptions clusterOptions, AdoNetStreamQueueMapper mapper, StreamingRelationalOrleansQueries queries, ILogger<AdoNetStreamFailureHandler> logger) : IStreamFailureHandler
 {
     private readonly ILogger<AdoNetStreamFailureHandler> _logger = logger;
