+++ test/Extensions/TesterAdoNet/Streaming/AdoNetQueueAdapterReceiverTests.cs
@@ -3,11 +3,9 @@ using MySql.Data.MySqlClient;
 using Orleans.AdoNet.Core;
 using Orleans.Configuration;
 using Orleans.Streaming.AdoNet;
-using Orleans.Tests.SqlUtils;
 using TestExtensions;
 using UnitTests.General;
 using static System.String;
-using RelationalOrleansQueries = Orleans.Streaming.AdoNet.Storage.RelationalOrleansQueries;
 
 namespace Tester.AdoNet.Streaming;
 
@@ -46,7 +44,7 @@ public abstract class AdoNetQueueAdapterReceiverTests(string invariant, TestEnvi
     private readonly TestEnvironmentFixture _fixture = fixture;
     private RelationalStorageForTesting _testing;
     private IRelationalStorage _storage;
-    private RelationalOrleansQueries _queries;
+    private StreamingRelationalOrleansQueries _queries;
 
     private const string TestDatabaseName = "OrleansStreamTest";
 
@@ -56,7 +54,7 @@ public abstract class AdoNetQueueAdapterReceiverTests(string invariant, TestEnvi
         Skip.If(IsNullOrEmpty(_testing.CurrentConnectionString), $"Database '{TestDatabaseName}' not initialized");
 
         _storage = _testing.Storage;
-        _queries = await RelationalOrleansQueries.CreateInstance(invariant, _storage.ConnectionString);
+        _queries = await StreamingRelationalOrleansQueries.CreateInstance(invariant, _storage.ConnectionString);
     }
 
     /// <summary>
