+++ test/Extensions/TesterAdoNet/Streaming/AdoNetQueueAdapterTests.cs
@@ -4,11 +4,9 @@ using Orleans.AdoNet.Core;
 using Orleans.Configuration;
 using Orleans.Streaming.AdoNet;
 using Orleans.Streams;
-using Orleans.Tests.SqlUtils;
 using TestExtensions;
 using UnitTests.General;
 using static System.String;
-using RelationalOrleansQueries = Orleans.Streaming.AdoNet.Storage.RelationalOrleansQueries;
 
 namespace Tester.AdoNet.Streaming;
 
@@ -47,7 +45,7 @@ public abstract class AdoNetQueueAdapterTests(string invariant, TestEnvironmentF
     private readonly TestEnvironmentFixture _fixture = fixture;
     private RelationalStorageForTesting _testing;
     private IRelationalStorage _storage;
-    private RelationalOrleansQueries _queries;
+    private StreamingRelationalOrleansQueries _queries;
 
     private const string TestDatabaseName = "OrleansStreamTest";
 
@@ -58,7 +56,7 @@ public abstract class AdoNetQueueAdapterTests(string invariant, TestEnvironmentF
         Skip.If(IsNullOrEmpty(_testing.CurrentConnectionString), $"Database '{TestDatabaseName}' not initialized");
 
         _storage = _testing.Storage;
-        _queries = await RelationalOrleansQueries.CreateInstance(invariant, _testing.CurrentConnectionString);
+        _queries = await StreamingRelationalOrleansQueries.CreateInstance(invariant, _testing.CurrentConnectionString);
     }
 
     /// <summary>
