+++ test/Extensions/TesterAdoNet/Streaming/AdoNetStreamFailureHandlerTests.cs
@@ -5,10 +5,8 @@ using Orleans.Configuration;
 using Orleans.Providers.Streams.Common;
 using Orleans.Streaming.AdoNet;
 using Orleans.Streams;
-using Orleans.Tests.SqlUtils;
 using UnitTests.General;
 using static System.String;
-using RelationalOrleansQueries = Orleans.Streaming.AdoNet.Storage.RelationalOrleansQueries;
 
 namespace Tester.AdoNet.Streaming;
 
@@ -45,7 +43,7 @@ public abstract class AdoNetStreamFailureHandlerTests(string invariant) : IAsync
 {
     private RelationalStorageForTesting _testing;
     private IRelationalStorage _storage;
-    private RelationalOrleansQueries _queries;
+    private StreamingRelationalOrleansQueries _queries;
 
     private const string TestDatabaseName = "OrleansStreamTest";
 
@@ -56,7 +54,7 @@ public abstract class AdoNetStreamFailureHandlerTests(string invariant) : IAsync
         Skip.If(IsNullOrEmpty(_testing.CurrentConnectionString), $"Database '{TestDatabaseName}' not initialized");
 
         _storage = _testing.Storage;
-        _queries = await RelationalOrleansQueries.CreateInstance(invariant, _testing.CurrentConnectionString);
+        _queries = await StreamingRelationalOrleansQueries.CreateInstance(invariant, _testing.CurrentConnectionString);
     }
 
     /// <summary>
