+++ test/Extensions/TesterAdoNet/Streaming/RelationalOrleansQueriesTests.cs
@@ -48,7 +48,7 @@ public abstract class RelationalOrleansQueriesTests(string invariant, int concur
     private const string TestDatabaseName = "OrleansStreamTest";
 
     private IRelationalStorage _storage;
-    private RelationalOrleansQueries _queries;
+    private StreamingRelationalOrleansQueries _queries;
 
     public async Task InitializeAsync()
     {
@@ -57,7 +57,7 @@ public abstract class RelationalOrleansQueriesTests(string invariant, int concur
 
         _storage = RelationalStorage.CreateInstance(invariant, testing.CurrentConnectionString);
 
-        _queries = await RelationalOrleansQueries.CreateInstance(invariant, testing.CurrentConnectionString);
+        _queries = await StreamingRelationalOrleansQueries.CreateInstance(invariant, testing.CurrentConnectionString);
     }
 
     private static string RandomServiceId(int max = 10) => $"ServiceId{Random.Shared.Next(max)}";
